src/recommendationservice/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.2.0         # via google-auth
certifi==2020.12.5        # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.26.3  # via -r requirements.in, google-api-python-client, google-cloud-core, google-cloud-monitoring, google-cloud-trace, opencensus
google-api-python-client==1.12.8  # via google-python-cloud-debugger
google-auth-httplib2==0.0.4  # via google-api-python-client, google-python-cloud-debugger
google-auth==1.24.0       # via google-api-core, google-api-python-client, google-auth-httplib2, google-python-cloud-debugger
google-cloud-core==1.5.0  # via google-cloud-trace
google-cloud-monitoring==0.36.0  # via opencensus-ext-stackdriver
google-cloud-profiler==1.1.2  # via -r requirements.in
google-cloud-trace==0.24.0  # via opencensus-ext-stackdriver
google-python-cloud-debugger==2.15  # via -r requirements.in
googleapis-common-protos==1.52.0  # via google-api-core
grpcio-health-checking==1.33.2  # via -r requirements.in
grpcio==1.33.2            # via -r requirements.in, google-api-core, grpcio-health-checking, opencensus-ext-grpc
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests

protobuf==3.13.0          # via google-api-core, googleapis-common-protos, grpcio-health-checking
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-json-logger==0.1.11  # via -r requirements.in
pytz==2018.9              # via google-api-core
pyyaml==5.3.1             # via google-python-cloud-debugger
requests==2.24.0          # via -r requirements.in, google-api-core
rsa==4.6                  # via google-auth
six==1.15.0               # via google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-cloud-core, google-python-cloud-debugger, grpcio, protobuf
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.10          # via requests
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry.sdk._logs a
opentelemetry-exporter-otlp-proto-grpc==1.10.0
opentelemetry-instrumentation-grpc==0.29b0
opentelemetry.instrumentation==0.29b0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
